requirements.txt
python-telegram-bot==23.14.1
fastapi==0.109.0
uvicorn==0.23.2
